Add p parallel-jobs option for make --jobs option
Add support for make using parallel jobs.
If this new option is not specified make is sequential.
If a value X is given to -p[X] or --parallel-jobs[=X] then X is passed
to make as the --jobs=X option.
If no value is passed, ie -p or --parallel-jobs is used, then the number
of parallel jobs will default to the number of processors.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>