Add ability to run configuration script
It is useful to be able to change certain default values, such as
DEF_BUF_SIZE, when using this script. Add a CLO to invoke a fixed
configuration script called configurable-options.sh in the project root
directory. Replacement values are then passed to configure to distribute
to Makefile xxxxx_CPPFLAGS.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>